web_dev/django/full_drf_tutorial/tutorial/requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
defusedxml==0.7.1
Django==4.1.7
django-filter==22.1
django-guardian==2.4.0
djangorestframework==3.14.0
djlint==1.19.16
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
httpie==3.2.1
idna==3.4
jsbeautifier==1.14.7
Markdown==3.4.1
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
